+++ src/komm/__init__.py
@@ -64,6 +64,7 @@ from ._error_control_decoders.ExhaustiveSearchDecoder import ExhaustiveSearchDec
 from ._error_control_decoders.GaussianEliminationDecoder import (
     GaussianEliminationDecoder,
 )
+from ._error_control_decoders.PeelingDecoder import PeelingDecoder
 from ._error_control_decoders.ReedDecoder import ReedDecoder
 from ._error_control_decoders.SCDecoder import SCDecoder
 from ._error_control_decoders.SyndromeTableDecoder import SyndromeTableDecoder
@@ -193,6 +194,7 @@ __all__ = [
     "BerlekampDecoder",
     "ExhaustiveSearchDecoder",
     "GaussianEliminationDecoder",
+    "PeelingDecoder",
     "ReedDecoder",
     "SCDecoder",
     "SyndromeTableDecoder",